+++ cpp-api-client/src/cpprest-client/model/Tag.cpp
@@ -25,7 +25,7 @@ Tag::Tag()
 {
     m_CreatedAt = 0L;
     m_CreatedAtIsSet = false;
-    m_Id = 0;
+    m_Id = 0L;
     m_IdIsSet = false;
     m_Text = utility::conversions::to_string_t("");
     m_TextIsSet = false;
@@ -86,7 +86,7 @@ bool Tag::fromJson(const web::json::value& val)
         const web::json::value& fieldValue = val.at(utility::conversions::to_string_t("id"));
         if(!fieldValue.is_null())
         {
-            int32_t refVal_id;
+            int64_t refVal_id;
             ok &= ModelBase::fromJson(fieldValue, refVal_id);
             setId(refVal_id);
         }
@@ -156,7 +156,7 @@ bool Tag::fromMultiPart(std::shared_ptr<MultipartFormData> multipart, const util
     }
     if(multipart->hasContent(utility::conversions::to_string_t("id")))
     {
-        int32_t refVal_id;
+        int64_t refVal_id;
         ok &= ModelBase::fromHttpContent(multipart->getContent(utility::conversions::to_string_t("id")), refVal_id );
         setId(refVal_id);
     }
@@ -195,12 +195,12 @@ void Tag::unsetCreatedAt()
 {
     m_CreatedAtIsSet = false;
 }
-int32_t Tag::getId() const
+int64_t Tag::getId() const
 {
     return m_Id;
 }
 
-void Tag::setId(int32_t value)
+void Tag::setId(int64_t value)
 {
     m_Id = value;
     m_IdIsSet = true;
